增加接口接受类 加密类

+++ src/main/java/cn/gyyx/sentinel/app/domain/Result.java
@@ -11,7 +11,7 @@ package cn.gyyx.sentinel.app.domain;
 public class Result<T> {
 
     // 错误码
-    private Integer code;
+    private Double code;
 
     // 提示信息
     private String msg;
@@ -19,11 +19,11 @@ public class Result<T> {
     // 具体的内容
     private T data;
 
-    public Integer getCode() {
+    public Double getCode() {
         return code;
     }
 
-    public void setCode(Integer code) {
+    public void setCode(Double code) {
         this.code = code;
     }
